+++ cryptofeed/defines.py
@@ -1,5 +1,5 @@
 '''
-Copyright (C) 2017-2023 Bryant Moscon - bmoscon@gmail.com
+Copyright (C) 2017-2024 Bryant Moscon - bmoscon@gmail.com
 
 Please see the LICENSE file for the terms and conditions
 associated with this software.
@@ -24,7 +24,6 @@ BITDOTCOM = 'BIT.COM'
 BITFLYER = 'BITFLYER'
 BITGET = 'BITGET'
 BITSTAMP = 'BITSTAMP'
-BITTREX = 'BITTREX'
 BLOCKCHAIN = 'BLOCKCHAIN'
 BYBIT = 'BYBIT'
 COINBASE = 'COINBASE'