+++ tools/generate_test_data.py
@@ -1,5 +1,5 @@
 '''
-Copyright (C) 2017-2023 Bryant Moscon - bmoscon@gmail.com
+Copyright (C) 2017-2024 Bryant Moscon - bmoscon@gmail.com
 
 Please see the LICENSE file for the terms and conditions
 associated with this software.
@@ -82,4 +82,4 @@ def main(only_exchange=None):
 
 
 if __name__ == '__main__':
-    main('OKCOIN')
+    main('BIT.COM')